+++ tests/integration/selectquerybuilder/test.php
@@ -24,3 +24,8 @@ $sqb->option( 'HAVING', [ 'safe' => $_GET['a'] ] ); // Safe
 $sqb->option( 'USE INDEX', $_GET['a'] ); // SQLi
 $sqb->option( 'IGNORE INDEX', $_GET['a'] ); // SQLi
 $sqb->option( 'SOME OTHER OPTION', $_GET['a'] ); // Safe
+
+$sqb->joinConds( [ 'sometable' => [ $_GET['join_type'], '1=1' ] ] ); // SQLi
+$sqb->joinConds( [ 'sometable' => [ 'LEFT JOIN', $_GET['a'] ] ] ); // SQLi
+$sqb->joinConds( [ 'sometable' => [ 'LEFT JOIN', [ $_GET['a'] ] ] ] ); // SQLi
+$sqb->joinConds( [ 'sometable' => [ 'LEFT JOIN', [ 'escaped' => $_GET['a'] ] ] ] ); // Safe