+++ Changelog.md
@@ -6,6 +6,7 @@
 
 ### New features
 * (MW) Check SQL `$options` passed to `SelectQueryBuilder::options` and `::option`
+* (MW) Check SQL `$join_conds` passed to `SelectQueryBuilder::joinConds`
 
 ### Bug fixes
 * (MW) Fix SQLi detection for `$options` and `$join_conds` arguments to `SELECT` DB functions